added jquery to enable autosizing the height of a text area
