+++ views/layouts/main.handlebars
@@ -65,7 +65,8 @@
 		<footer>
 
 		</footer>
-
+		<script type="module" src="/js/script.js"></script>
+		<script src="https://code.jquery.com/jquery-3.6.3.slim.min.js" integrity="sha256-ZwqZIVdD3iXNyGHbSYdsmWP//UBokj2FHAxKuSBKDSo=" crossorigin="anonymous"></script>
 		{{#if logged_in}}
 			<script src="/js/logout.js"></script>
 		{{else}}
